Split test_decoders.py into one file per decoder
The single TestEncodings class held 25 tests across four encodings, and ZRLE, Tight and Cursor are still to come. Each decoder now has its own module over a shared tests/unit/decoder_fixtures.py, so the next encoding lands in a new file rather than growing a common one.

The fixture module keeps only what more than one decoder needs; the three hextile_* builders stay in test_decoder_hextile.py, next to their only caller. CoRRE's tests sit beside RRE's because the decoder is a subclass and the two sets are line-for-line parallel -- a divergence between them is worth seeing in one file.

Test bodies are unchanged; only the enclosing class names differ, since a module named for its decoder makes TestEncodings the wrong name. Imports are absolute (tests.unit.decoder_fixtures) so a module resolves both under `discover tests/unit` and as `python -m unittest tests.unit.test_...`.

302 tests before, 302 after.
